+++ repos/base-hw/src/base/signal/signal.cc
@@ -150,7 +150,7 @@ Signal_context_capability Signal_receiver::manage(Signal_context * const c)
 	Signal_connection * const s = signal_connection();
 	while (1) {
 		try {
-			c->_cap = s->alloc_context(_cap, (unsigned)c);
+			c->_cap = s->alloc_context(_cap, (unsigned long)c);
 			c->_receiver = this;
 			_contexts.insert(&c->_receiver_le);
 			return c->_cap;
